backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==4.2.11
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==3.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==25.0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
python-dotenv==1.0.0